Fix illegal VHDL in csr_pkg.vhd
It is not legal VHDL to have a port mode in a record.
I have developed a VHDL-parser and language server (https://github.com/kraigher/rust_hdl/) and was using you repo as a test when I found this error. 
The parser can process all VHDL files in your repo in 100 ms.